+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [1.1.0] - 2020-06-13
+### Changed
+- Changed default MTU size to 20.
+
+### Fixed 
+- Brought in missing <string> include in NrfDfuServer.cpp
+- Misc formatting and spelling
 
 ## [1.0.1] - 2020-08-17
 ### Fixed